refractor

+++ src/main/java/ru/test/Downloader/DownloaderImpl.java
@@ -3,12 +3,15 @@ package ru.test.Downloader;
 import com.google.common.util.concurrent.RateLimiter;
 import org.springframework.stereotype.Component;
 
+import java.io.BufferedInputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.net.URL;
 import java.nio.channels.Channels;
 import java.nio.channels.ReadableByteChannel;
+import java.nio.file.Files;
+import java.nio.file.Path;
 
 @Component
 public class DownloaderImpl implements Downloader {
@@ -19,15 +22,35 @@ public class DownloaderImpl implements Downloader {
      */
     @Override
     public void download(String url, String filePath, double rateLimit) throws IOException {
-        RateLimiter limiter = RateLimiter.create(rateLimit);
-        URL website = new URL(url);
-        ReadableByteChannel rbc = Channels.newChannel(website.openStream());
-
+        BufferedInputStream bis = null;
+        FileOutputStream fos = null;
+        ThrottledOutputStream tos = null;
+        if (!Files.isDirectory(Path.of(filePath))) {
+            Files.createDirectory(Path.of(filePath));
+        }
         File file = new File(filePath, url.substring(url.lastIndexOf("/")));
+        if (!file.exists()) {
+            try {
+                bis = new BufferedInputStream(new URL(url).openStream());
+                fos = new FileOutputStream(file);
+                tos = new ThrottledOutputStream(fos, rateLimit);
 
-        FileOutputStream fos = new FileOutputStream(file);
-
-        limiter.acquire();
-        fos.getChannel().transferFrom(rbc, 0, Long.MAX_VALUE);
+                final byte[] bytes = new byte[1024];
+                int count;
+                while ((count = bis.read(bytes)) != -1) {
+                    tos.write(bytes, 0, count);
+                }
+            } finally {
+                if (bis != null) {
+                    bis.close();
+                }
+                if (fos != null) {
+                    fos.close();
+                }
+                if (tos != null) {
+                    tos.close();
+                }
+            }
+        }
     }
 }
